// types/hotel.ts
export interface HotelTransferData {
    name: string;
    shortName: string;
    cityName: string;
    area: string;
    complex: string;
    distanceKm: number;
    driveTimeMin: number;
    haramAccess: string;
    haramName: string;
    highlight: string;
    dropOffNote: string;
    hubUrl: string;
    hubLabel: string;
}